midterms in view all section. restored all cards and updated api endpoint

+++ lib/core/models/class_schedule_model.dart
@@ -41,7 +41,7 @@ class ClassData {
   String termCode;
   String subjectCode;
   String courseCode;
-  int units;
+  double units;
   String courseLevel;
   String gradeOption;
   String grade;
